layout: Use the background's surface ivi_output when activating apps by default
In order to activate application by default, we need a ivi_output which
for the desktop role that's not available at that time (the activation
part actually will set-up one). Uses the bg output for this case.

This is only activated by adding 'activate-by-default' bool variable
under the '[shell]' section (in the ini configuration file).

Bug-AGL: SPEC-3118

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ib44f0fcccc145216fb28fc9c26e5a065912ceef5