shell: Add a black surface in the fullscreen layer
Now that we're capable of restarting the client shell without the need
to restart the compositor, create a black surface and insert in the
fullscreen layer as to denote that the client shell is no longer
running. This black surface is removed when the 'ready' request is
received and inserted back when the client shell unbinds from the
agl-shell protocol.

Also, we were missing implementation protocol specification as the
presentation delay required a black surface being displayed instead, so
this brings the implementation closer to that of the protocol
specification.

Bug-SPEC: SPEC-3161

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I40f01135583eea8af78d3077cdad97ad5ad450f5