meson.build: Build X11 backend optionally
libweston6 built and installed the x11 backend unconditionally, but this
is no longer the case with newer version of weston.  Also, poky-agl
configuration specifically removes any distro features that contain x11.

Users wanting to test the compositor under a x11 backend need to
compile their own libweston version this shouldn't be a problem.

Unfortunatelly, meson does not have a easy way to determine if a header
is installed or not so after glancing over the documentation we need to
use the compilers' 'has_header' function to determine if the x11 header
was installed or not by the weston recipe.

This makes sure that if we're building out-of-tree it finds the proper
header as well.

Bug-AGL: SPEC-3180

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I3dce7c1beaf5bfab31fb88bc0462c7eb8320cfc8