src/ Adding remote surface role
This patch adds the 'remote' surface role, which clients can make use of
to hint the compositor that is should place the surface on other
output.

While both private extension protocols explicitly require to provide an
wl_output when activating or when setting surface roles, we still need
the inform the compositor that the surface should be placed on another
output. This is due to the activate_by_default functionality that
requires having an output being present, with the default regular XDG
desktop roles deriving its own output by using output of the backgound
surface role.

Just like pop-up dialog role and split surface role this patch adds
another temporary hold up place before the surface is actually created.
Once that is done the surface role will be re-assigned to the desktop
role (being no difference whatsoever between them).

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I8804cc3478974f0b0cb323db46cb0814405c34d1