src/: Add the remote surface role
This patch adds the 'remote' surface role, which clients can make use of
to hint the compositor that it should place the surface on other
output.

While both private extension protocols (agl-shell and agl-shell-desktop)
explicitly  require a wl_output when activating or when setting surface
roles, we still need the inform the compositor that the surface should be
placed on another output.

This is due to the activate_by_default functionality that requires
having an output being present, with the default regular XDG desktop
role deriving its output by using the output of the backgound surface
(which is being normally set by the client shell).

Just like pop-up dialog role and split surface role this patch adds
another temporary hold up place before the surface is actually created.

Bug-AGL: SPEC-3280

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ic67246ecc183826ae59b2c99a671885d61040249