main: Enable remote outputs later than local (DRM) outputs
Will allow the client shell to receive, and to use as primary output, as
the local outputs first without the need to (pre-)configure the client
shell or the other applications -- making all of the assume the remote
output the last sent by the xdg_output protocol. Makes integration much
easier and will retain the same functionality we had with weston.

Bug-AGL: SPEC-3280

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Idc91383aa6edb5f859ead43246906451ed08cd2d