layout: Clip the view to the size specified by width and height
This uses the bx and by as origin for the bounding box if those were
passed on. The compositor will set up a mask clip for the window if the
both width and height are positive.

This also removes the geometry of the surface and passes the x and y values
directly.

Bug-AGL: SPEC-3419

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I174f1ccbffad3c99e74fbb5ee90888b4ad879230