layout: Activate by default xdg surface that do not have an app_id set
The following is necessary in case the app_id is set-up at a later point
in time, and it is intended as a work-around for WAM/chromium68 clients
that specify the app_id after the initial surface commit. Note that
activation will not work, in case the application will never set-up an
app_id because we can't identify which app_id to switch to.

The activation phase is needed to inform the client of the window
dimensions, and it is mandated by the XDG-shell protocol.

Bug-AGL: SPEC-3544

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I49a7270dcb68e525fc9d85db2cd189a50c02502f