shell: Advertise app when app_id is available
When using GTK-3 to create apps, the framework issues set_parent_id
which forces a surface to be added even before the app_id is set.
This causes the compositor to fail to advertise the app.

This patch tries to overcome this issue by checking on surface commit
if the surface was previously advertised, and do it if it was not.

Signed-off-by: Walter Lozano <walter.lozano@collabora.com>
Change-Id: I4e5fec82c1008c30d8ae7d213d85a82e39c404c3