main: Add support loading waltham transmitter plug-in
This is the same as the remoting plug-in, given that we'll be using the
remoting plug-in to create the remoting output.

The difference would be that instead of the 'remoting-output' section
we'll use 'transmitter-output', with all other section specific being
the same. For instance, using agl-shell-app-id we can instruct the
application to be assigned to that 'transmitter-output'.

Bug-AGL: SPEC-3611

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ic16b6b9672802f4f83be121385c0798b9dbe08e9