shell: Push the surface to the remote side
The receiver expects that the sender/transmitter will inform it when it
can create a surface and it can start the gstreamer pipeline.

The transmitter plug-in exposes a way to call that using 'push_surface'
callback. So, whenever we detect that the remote surface is to be
assigned to a waltham-type of output, we would need to call into this
'push_surface' callback. This acts as notification mechanism for the
remote side.

Bug-AGL: 3611

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ia5db0d8f2356708eaee3fbaabed2b3ba3ac97eb6