desktop: Delay surface creation until committed
Some frameworks like GTK-3 delay the set_app_id until title and
parent are set, which makes ivi_check_pending_desktop_surface
unable to set the proper role on surface_added. A consequence
of this behaviour is that is not possible to use set_app_property
to configure an application as popup.

Instead, delay the creation of the surface until committed to give
the chance to have a valid app_id configured.

Signed-off-by: Walter Lozano <walter.lozano@collabora.com>
Change-Id: I3f672fb365f48659711c866c45a21df6058a5a02