desktop: Allow to remove surface from pending list
In some special corner cases a surface can be removed while it is
still in the pending list. An example of this case is when a surface
is added and removed before the system is ready. Under this circumstances
the surface still has role == ROLE_NONE and is in the pending list.

The fact of not removing the surface from the pending_list ends in a
segfault when the system is ready.

In order to fix the issue, remove the restriction to only remove surfaces
with role != ROLE_NONE.

Signed-off-by: Walter Lozano <walter.lozano@collabora.com>
Change-Id: Ia71b6c3ca214849448c935cc09039581d0e209a8