meson.build: Don't set the dependency type for wayland-scanner
meson (0.53) seem to be smart enough to resolve cross-compile
dependencies without setting explicitly the dependency type.

Setting the dependency  as native results in ignoring PKG_CONFIG_PATH
and using PATH to choose and alternative pkg-config. Sourcing the
yocto/OE toolchain would result in picking the hosts pkg-config due to
modified a PATH where SDKTARGETSYSROOT is before any of the hosts paths.

Combined that with the fact that PKG_CONFIG_PATH is reset when the
dependency is se to native (why that hapens is something to further
investigate) results in not finding wayland-scanner.

Tested with both yocto/OE toolchains and native (local) builds to make
sure we can still build it and seems we're fine.

Bug-AGL: SPEC-4169

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I5c4b12eff044aeb0c3ab4c18e943da4e15d6c411