agl-shell: Add bound_ok and bound_fail events and a default destructor
As we're increasing the amount of combinations we could have in the AGL
platforms, in which we're mixing various platforms, we require a way
to tell clients that there's already a shell client which is charge.

We can't really have multiple shell clients managing the windows and
surfaces but in the same time we don't want to sever the waylad
connection the way it happens in the first version of this protocol
interface. Racy alternatives might exist, which basically avoids
advertising the interface altogether once a client did bind to the
interface but that's not really a viable solution.

So, instead of doing that, this patch introduces two new events which
tells either client that it was either successful, and it can continue
issue requests and receive events and a fail event.

While doing an update to protocol, to add those two events add also
a default destructor for the protocol.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>