layout: Detect when an application should be migrated to other outputs
This patch adds a way to determine when to move out/migrate the window to another
output, different than one set previously, or where it was initially added.

This patch only adds support for resizing with a later patch touch the
final activation part.

Uses new output variable to track it the outputs. We send new dimensions
of the output if they do not match the previous/older recorded one.

Bug-AGL: SPEC-4516
Reported-by: Vasyl Vavrychuk <vasyl.vavrychuk@opensynergy.com>
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ic5b6b9c1676543fa5ff1524f58e014ad8abb632b