clients/grpc: Initial start for gRPC proxy
The proxy helper client would bind to the agl-shell protocol interface
but also create a gRPC server which implements the agl-shell protobuf
interface.

The gRPC server implementation created in this helper client would be
used by regular clients if they would require further change the window
management or change window properties.

Note that this is an initial bring-up with some of implementation being
a stub, as  the window properties would actually be implemented when
expanding the agl-shell private extension.

The hooks in the gRPC implementation that are in place are:

- xxx
- yyy
- zzz
- ttt

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I5f5e8426d52ed7bf2e264be78c6572388afa53af